+++ centimeter/components/Presentation/SlideTypes.vue/ContentSlides.vue
@@ -3,7 +3,7 @@
         <div class="flex items-center gap-2 mb-3">
             <h3 class="text-sm font-medium text-[var(--faded-text-color)]">Content slides</h3>
             <button class="text-[var(--faded-text-color)] hover:text-[var(--text-color)]">
-            <HelpCircle :size="16" />
+            <HelpCircle :size="16" @click.stop="onIconClick" />
             </button>
         </div>
         <div class="grid grid-cols-2 gap-3">
@@ -13,7 +13,7 @@
             class="flex items-center gap-3 p-3 rounded-lg border border-[var(--faded-bg-color)] hover:border-[var(--faded-bg-color-dark)] hover:bg-[var(--faded-bg-color)] transition-all text-left"
             @click="handleClick(option)"
             >
-            <component :is="option.icon" :size="20" :class="option.color" />
+            <component :is="option.icon" :size="20" :class="option.color" @click.stop="onIconClick" />
             <span class="text-sm font-medium text-[var(--text-color)]">{{ option.label }}</span>
             </button>
         </div>
@@ -22,11 +22,16 @@
 
 <script setup lang="ts">
 import {Type, Image, Play, LayoutGrid} from 'lucide-vue-next'
-import type {Option} from '../utils/types';
+import type {Option} from "../../../utils/types.ts"
 
 function handleClick(option: Option): void {
   console.log('Option clicked:', option.label)
   emits('add-slide', option.label)
+  emits('toggle-icon')
+}
+
+function onIconClick(): void {
+  emits('toggle-icon')
 }
 
 const contentOptions = [
@@ -37,6 +42,7 @@ const contentOptions = [
 ]
 
 const emits = defineEmits<{
-  'add-slide': [slideType: string]
+  'add-slide': [slideType: string],
+  'toggle-icon': []
 }>()
 </script>
